pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "indico-mcp"
version = "0.1.0"
description = "MCP server for the Indico meeting agenda system"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.scripts]
indico-mcp = "indico_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/indico_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]
